// libs/components/src/lib/bracket/drawing/grid/types.ts
import { BracketDataLayout } from '../../core';
import { FinalizedBracketColumn, FinalizedBracketMatchElementMap, MutableBracketGrid } from './core';

export type ComputedBracketGrid<TRoundData, TMatchData> = {
  raw: MutableBracketGrid<TRoundData, TMatchData>;
  columns: FinalizedBracketColumn<TRoundData, TMatchData>[];
  matchElementMap: FinalizedBracketMatchElementMap<TRoundData, TMatchData>;
};

export type CreateBracketGridConfig = {
  includeRoundHeaders: boolean;
  columnWidth: number;
  matchHeight: number;
  roundHeaderHeight: number;
  roundHeaderGap: number;
  columnGap: number;
  rowRoundGap: number;
  rowGap: number;
  finalMatchHeight: number;
  finalColumnWidth: number;
  swissGroupPadding: number;
  swissGroupBorderWidth: number;
  layout: BracketDataLayout;
  continueElement?: {
    columnWidth: number;
    elementHeight: number;
  } | null;
};